Faculty reject button is fully operational

+++ TutorMaster/TutorMaster/FacultyMain.cs
@@ -17,37 +17,40 @@ namespace TutorMaster
         {
             id = accID;
             InitializeComponent();
-            SetupPendingRequests(id);//Populate box with requests, similar to Admin main
+            SetupPendingRequests(id);                                                                                       //Populate box with requests, similar to Admin main
         }
 
-        public void SetupPendingRequests(int accID) //This function will populate the checked list box with pending requests
+        public void SetupPendingRequests(int accID)                                                                         //This function will populate the checked list box with pending requests
         {
-            //EDIT THIS TO MAKE SOME COLUMNS SO THAT I CAN QUERY INFORMATION LATER
 
             lvPendingRequests.CheckBoxes = true;
-            //1. find which classes the faculty members are qualified to approve tutors for
-            TutorMasterDBEntities4 db = new TutorMasterDBEntities4(); //create a new indirect entity to look at db
+            lvPendingRequests.Columns.Add("Student Name", 100);
+            lvPendingRequests.Columns.Add("Class Code", 100);
+            lvPendingRequests.Columns.Add("ID", 100);
 
-            var ApprovableClasses = (from row in db.FacultyClasses where row.FacultyID == accID select row.ClassCode); //This should pull out all the classes that the faculty member can approve tutors for
-            List<String> AC= new List<String>(); //create a list to store the DB info that was just pulled
-            AC= ApprovableClasses.ToList<String>(); //Fill the list
+            //1. find which classes the faculty members are qualified to approve tutors for
+            TutorMasterDBEntities4 db = new TutorMasterDBEntities4();                                                       //create a new indirect entity to look at db
+
+            var ApprovableClasses = (from row in db.FacultyClasses where row.FacultyID == accID select row.ClassCode);      //This should pull out all the classes that the faculty member can approve tutors for
+            List<String> AC= new List<String>();                                                                            //create a list to store the DB info that was just pulled
+            AC= ApprovableClasses.ToList<String>();                                                                         //Fill the list
             
-            var PendingRequests = (from row in db.TutorRequests select row); //This pulls out the whole table of pending requests
-            List<TutorRequest> PC = new List<TutorRequest>(); //create a list to store the DB info that was just pulled
-            PC = PendingRequests.ToList<TutorRequest>(); //Fill the list
+            var PendingRequests = (from row in db.TutorRequests select row);                                                //This pulls out the whole table of pending requests
+            List<TutorRequest> PC = new List<TutorRequest>();                                                               //create a list to store the DB info that was just pulled
+            PC = PendingRequests.ToList<TutorRequest>();                                                                    //Fill the list
 
             //1.5 Create an empty list to fill with the right requests in step 2
-            List<TutorRequest> FC = new List<TutorRequest>(); //here is that beautiful empty list that will be filled with the tutor requests that should be displayed
+            List<TutorRequest> FC = new List<TutorRequest>();                                                               //here is that beautiful empty list that will be filled with the tutor requests that should be displayed
 
             //2. Look at all pending requests and see if they are for a class from the faculty members department. If they are, add them to the box
             
-            foreach(String classcode in AC) //Iterate through all of the class  codes and check them
+            foreach(String classcode in AC)                                                                                 //Iterate through all of the class  codes and check them
             {
-                foreach (TutorRequest tut in PC) // Iterate through all of the requests
+                foreach (TutorRequest tut in PC)                                                                            // Iterate through all of the requests
                 {
-                    if (classcode == tut.ClassCode) // If the class code that we are checking matches the class code of the object,
+                    if (classcode == tut.ClassCode)                                                                         // If the class code that we are checking matches the class code of the object,
                     {
-                        FC.Add(tut); // Then add that tutor request to the list of requests to display
+                        FC.Add(tut);                                                                                        // Then add that tutor request to the list of requests to display
                     }
                 }
             }
@@ -55,21 +58,21 @@ namespace TutorMaster
             //3. Now that all the requests are in the FC, we need to display each of the tutor requests in clbPendingReequests in the form of [STUDENT NAME, CLASS TO TUTOR]
             //3.5 I have to go through the student table and pull out the names of the people in it
             
-            var students = from c in db.Users select c; // This creates a list of Users that I can use to pull information from 
+            var students = from c in db.Users select c;                                                                     // This creates a list of Users that I can use to pull information from 
             List<User> stus = new List<User>();
-            stus = students.ToList(); //The list 'stus' includes all of the users from the users db
+            stus = students.ToList();                                                                                       //The list 'stus' includes all of the users from the users db
 
-            List<string> Names=new List<string>(); //The list Names is empty, but will be used to contain all the Names that correspond to the proper requests
+            List<string> Names=new List<string>();                                                                          //The list Names is empty, but will be used to contain all the Names that correspond to the proper requests
             string Name = "";
 
-            foreach (TutorRequest req in FC)//iterate through all the requests to display
+            foreach (TutorRequest req in FC)                                                                                //iterate through all the requests to display
             {
-                foreach (User stu in stus)//iterate through all the users [probably not the most effecient way to do this]
+                foreach (User stu in stus)                                                                                  //iterate through all the users [probably not the most effecient way to do this]
                 {
-                    if (req.ID == stu.ID) //When you find the matching ID
+                    if (req.ID == stu.ID)                                                                                   //When you find the matching ID
                     {
-                        Name = stu.FirstName + " " +stu.LastName; //Create the name string
-                        Names.Add(Name); //Add the Name to the list of names 
+                        Name = stu.FirstName + " " +stu.LastName;                                                           //Create the name string
+                        Names.Add(Name);                                                                                    //Add the Name to the list of names 
                     }
                 }
             }
@@ -77,7 +80,7 @@ namespace TutorMaster
             //4. Now that all the data has been prepared, display it properly
             for(int i =0; i<FC.Count();i++) //loop through the parrallel lists
             {
-                lvPendingRequests.Items.Add(Names.ElementAt(i)+" "+FC.ElementAt(i).ClassCode); //add the items as a string into the box!
+                lvPendingRequests.Items.Add(new ListViewItem(new string[] { Names.ElementAt(i), FC.ElementAt(i).ClassCode, FC.ElementAt(i).ID.ToString()})); //add the items as a string into the box!
             }
         }
 
@@ -85,7 +88,6 @@ namespace TutorMaster
         {
             //System.Windows.Forms.Application.Exit();
         }
-
 
         private void btnLogout_Click(object sender, EventArgs e)
         {
@@ -99,27 +101,36 @@ namespace TutorMaster
             //This will need to change each student account selected and then remove the pending requests from the DB
         }
 
-        private void btnReject_Click(object sender, EventArgs e) //In order to work around the issues for this function, I need to reformat the list view
+        private void btnReject_Click(object sender, EventArgs e)                                                    //This Button now works fully! addition functionality would be eror checking for parse function and sending message to students and administrators once we set that up. 
         {
             //This will need to remove all the requests from the DB and leave all acounts unchanged. Eventually, it will send a message to the admin account
 
             
-            TutorMasterDBEntities4 db = new TutorMasterDBEntities4(); //open the db
+            TutorMasterDBEntities4 db = new TutorMasterDBEntities4();    //side note, we should carefully control the life cycle of object contexts like these                                           //open the db
 
-            int reqNum = lvPendingRequests.CheckedItems.Count; //identify how many requests have been checked
-            for (int i = 0; i < reqNum; i++) //iterate for how many requests have been clicked
+            int reqNum = lvPendingRequests.CheckedItems.Count;                                                      //identify how many requests have been checked
+            for (int i = 0; i < reqNum; i++)                                                                        //iterate for how many requests have been clicked
             {
 
 
-                string request = lvPendingRequests.CheckedItems[i].Text; //pull the classcode from the request in question
-                //This line creates text objects that look like 'FIRST LAST CLASS-123' you may want to split the string to get the class code, but you also need the student ID
-                TutorRequest delU = (from row in db.TutorRequests where ((row.ClassCode == request) ) select row).First(); //find the request that has the right ID and class code
-                
-                
-                db.TutorRequests.DeleteObject(delU);//delete the object in the db
-                db.SaveChanges();//save the cahnges to the db
-            }
+                string CC = lvPendingRequests.CheckedItems[i].SubItems[1].Text;                                     //pull the classcode from the request in question
+                string Id = lvPendingRequests.CheckedItems[i].SubItems[2].Text;
+                int Id2 = Int32.Parse(Id);                                                                         //Change the string into an int
 
+
+                //This line creates text objects that look like 'FIRST LAST CLASS-123 ID' you may want to split the string to get the class code, but you also need the student ID
+
+                TutorRequest delU = (from row in db.TutorRequests where ((row.ClassCode == CC) && (row.ID == Id2)) select row).First(); //find the request that has the right ID and class code
+
+                if (delU != null)
+                {
+                    db.TutorRequests.DeleteObject(delU);                                                                //delete the object in the db
+                    db.SaveChanges();                                                                                   //save the cahnges to the db
+
+                }
+            }
+            lvPendingRequests.Clear(); //clean out the box
+            SetupPendingRequests(id); //Set up the box again
         }
 
         private void listView1_SelectedIndexChanged(object sender, EventArgs e)
